add Titlemenu

+++ Game/Game/source/MoveComponent.cpp
@@ -38,8 +38,8 @@ void MoveComponent::ProcessInput()
 	if (key & PAD_INPUT_RIGHT) { v.z = 1; }
 	*/
 	GetJoypadDirectInputState(DX_INPUT_KEY_PAD1 , &input);
-	v.x = input.Y / 1000;
-	v.z = input.X / 1000;
+	v.x = (float)input.Y / 1000;
+	v.z = (float)input.X / 1000;
 
 
 	// v‚ðrad•ª‰ñ“]‚³‚¹‚é
